requirements.txt
evdev==1.7.1
MouseInfo==0.1.3
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
six==1.16.0
